+++ src-ui/pages/level_viewer/tiles.ts
@@ -13,7 +13,7 @@ interface ImageSpec {
 
 function getTerrainFunc(alpha?: number) {
   return (ctx): ImageSpec[] => {
-    if (ctx.area == "Jungle") {
+    if (["Jungle", "Black Market"].includes(ctx.area)) {
       return [{ name: "alltiles", x: 512, y: 128, alpha: alpha }];
     } else if (ctx.area == "Worm") {
       return [{ name: "alltiles", x: 0, y: 17 * 64, alpha: alpha }];
@@ -28,7 +28,7 @@ function getTerrainFunc(alpha?: number) {
 
 function getTerrainSpikesFunc(alpha?: number) {
   return (ctx): ImageSpec[] => {
-    if (ctx.area == "Jungle") {
+    if (["Jungle", "Black Market"].includes(ctx.area)) {
       return [{ name: "alltiles", x: 64 * 13, y: 64 * 6, alpha: alpha }];
     } else if (ctx.area == "Worm") {
       return [{ name: "alltiles", x: 64 * 5, y: 64 * 22, alpha: alpha }];
@@ -102,7 +102,7 @@ export default {
   },
   L: {
     images: function (ctx) {
-      if (ctx.area == "Jungle") {
+      if (["Jungle", "Black Market"].includes(ctx.area)) {
         if (ctx.below != "L") {
           return [{ name: "alltiles", x: 576, y: 0 }];
         }
@@ -126,7 +126,7 @@ export default {
   },
   Q: {
     images: function (ctx) {
-      if (ctx.area == "Jungle") {
+      if (["Jungle", "Black Market"].includes(ctx.area)) {
         return [
           { name: "alltiles", x: 960, y: 192 },
           { name: "alltiles", x: 576, y: 0, offY: 64 },
@@ -229,6 +229,12 @@ export default {
   },
   c: {
     images: function (ctx) {
+      if (ctx.area == "Worm") {
+        return [
+          { name: "items", x: 80 * 11, y: 80 * 7, offY: -20, w: 80, h: 80 },
+          { name: "water", x: 64, y: 0 },
+        ];
+      }
       return [{ name: "items", x: 1040, y: 0, offX: 24, w: 80, h: 80 }];
     },
   },
@@ -315,7 +321,7 @@ export default {
       let x = 0;
       let y = 0;
 
-      if (ctx.area == "Jungle") {
+      if (["Jungle", "Black Market"].includes(ctx.area)) {
         x = 256 * 2;
       } else if (ctx.area == "Worm") {
         x = 256 * 2;
@@ -343,7 +349,7 @@ export default {
   },
   ":": {
     images: function (ctx) {
-      if (ctx.area == "Jungle") {
+      if (["Jungle", "Black Market"].includes(ctx.area)) {
         return [{ name: "monsters5", x: 0, y: 80 * 3, w: 80, h: 80 }];
       }
       return [{ name: "monsters5", x: 0, y: 0, w: 80, h: 80 }];
@@ -405,7 +411,7 @@ export default {
   // Has many variants...
   h: {
     images: function (ctx) {
-      if (ctx.area == "Jungle") {
+      if (["Jungle", "Black Market"].includes(ctx.area)) {
         return [{ name: "alltiles", x: 12 * 64, y: 5 * 64 }];
       }
       // Hell Bricks